Separate history related code to file
